+++ server/constructors_test.go
@@ -9,7 +9,7 @@ import (
 
 func Test_openDB(t *testing.T) {
 	t.Parallel()
-	_, err := openDB(t.TempDir())
+	_, err := openDB(t.TempDir(), "")
 	require.NoError(t, err)
 }