+++ server/mock/app.go
@@ -20,7 +20,7 @@ import (
 // similar to a real app. Make sure rootDir is empty before running the test,
 // in order to guarantee consistent results
 func NewApp(rootDir string, logger log.Logger) (abci.Application, error) {
-	db, err := sdk.NewLevelDB("mock", filepath.Join(rootDir, "data"))
+	db, err := sdk.NewDB("mock", "", filepath.Join(rootDir, "data"))
 	if err != nil {
 		return nil, err
 	}