+++ server/start.go
@@ -175,7 +175,7 @@ func startStandAlone(ctx *Context, appCreator types.AppCreator) error {
 	transport := ctx.Viper.GetString(flagTransport)
 	home := ctx.Viper.GetString(flags.FlagHome)
 
-	db, err := openDB(home)
+	db, err := openDB(home, ctx.Viper.GetString("db_backend"))
 	if err != nil {
 		return err
 	}
@@ -235,7 +235,7 @@ func startInProcess(ctx *Context, clientCtx client.Context, appCreator types.App
 	}
 
 	traceWriterFile := ctx.Viper.GetString(flagTraceStore)
-	db, err := openDB(home)
+	db, err := openDB(home, cfg.DBBackend)
 	if err != nil {
 		return err
 	}